requirements.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
openai==1.30.1
packaging==24.0
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tqdm==4.66.4
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.3
